refactor(data-layer): split the methods into multiple files
